Change how logical cores per package is read
Changed to adding + 1 in a spot, as it was needed to comply with the
AMD programmers manual, and avoid an off-by-one error on the 256-core
AMD EPYC processor.

This could be erroneous for other processors, and is not properly tested.
It does seemingly not apply to Inter processors, as they have a special
path in the code. Maybe we should have a similar other path for AMD.